Remove redundant package from preview install command
Qemu doesn't seem to be available standalone, so the command fails. It's already a part of qemu-system, so it's not necessary to include qemu itself anyway.

This is just the result I got when I tried to install on Debian though, might want to check with Ubuntu just in case.